fix IPv6 parsing, use toUShort instead of toInt

+++ ktor-network/posix/src/io/ktor/network/util/SocketAddressUtils.kt
@@ -54,10 +54,10 @@ internal fun parseIPv6String(ipString: String): ByteArray? {
 
         groups.flatMap {
             val int = if (it.matches(Regex("^[0-9a-fA-f]+$"))) {
-                it.toInt(16)
+                it.toUShort(16)
             } else {
                 return null
-            }
+            }.toInt()
             listOf((int shr 8).toByte(), int.toByte())
         }.toByteArray()
     } catch (_: Throwable) {
